add shared external package for b64 utils + readme
