+++ examples/shared/base64_utils.py
@@ -1,0 +1,17 @@
+import base64
+from io import BytesIO
+
+from PIL import Image
+
+BASE64_PREAMBLE = "data:image/png;base64,"
+
+
+def pil_to_b64(pil_img):
+    buffered = BytesIO()
+    pil_img.save(buffered, format="PNG")
+    img_str = base64.b64encode(buffered.getvalue())
+    return BASE64_PREAMBLE + str(img_str)[2:-1]
+
+
+def b64_to_pil(b64_str):
+    return Image.open(BytesIO(base64.b64decode(b64_str.replace(BASE64_PREAMBLE, ""))))
